requirements.txt
appdirs==1.4.3
attrs==19.3.0
Automat==0.8.0
autopep8==2.0.1
beautifulsoup4==4.11.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cloud-init==22.3.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docopt==0.6.2
entrypoints==0.3
filelock==3.8.2
Flask==2.2.2
Flask-PyMongo==2.3.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==5.1.0
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
kafka-python==2.0.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.1.1
more-itertools==4.2.0
msgpack==1.0.4
netifaces==0.10.4
newsapi-python==0.2.6
oauthlib==3.1.0
pexpect==4.6.0
pipreqs==0.4.11
platformdirs==2.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.10.0
pycountry==22.3.5
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
pymediawiki==0.7.2
pymongo==4.3.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
pyxattr==0.6.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
soupsieve==2.3.2.post1
ssh-import-id==5.10
systemd-python==234
tomli==2.0.1
Twisted==18.9.0
ubuntu-advantage-tools==27.11.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==2.2.2
yarg==0.1.9
youtube-dl==2020.3.24
zipp==1.0.0
zope.interface==4.7.1
